weather_application_backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
fonttools==4.51.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
